fix: 优化定时器功能和音效管理逻辑
- 在 GlobalPlayerService 中安全地获取音效列表，处理可能的异常情况，确保应用稳定性。
- 修改 TimerDialog 中的定时器选择逻辑，支持浮点数分钟设置，提升用户体验。
- 更新定时器设置的 Snackbar 消息，确保显示正确的分钟数。

+++ lib/shared/widgets/timer_dialog.dart
@@ -88,8 +88,6 @@ class TimerDialog extends StatelessWidget {
                   onTimerSet?.call();
                 },
               ),
-
-
             ],
           ),
         ),
@@ -136,8 +134,11 @@ class TimerDialog extends StatelessWidget {
           children: [5, 10, 15, 30, 45, 60].map((minutes) {
             final isSelected = currentTimerMinutes == minutes;
             return InkWell(
-              onTap: () =>
-                  _handleTimerSelection(context, minutes, localizations),
+              onTap: () => _handleTimerSelection(
+                context,
+                minutes.toDouble(),
+                localizations,
+              ),
               borderRadius: BorderRadius.circular(20),
               child: Container(
                 padding: const EdgeInsets.symmetric(
@@ -189,16 +190,16 @@ class TimerDialog extends StatelessWidget {
 
   void _handleTimerSelection(
     BuildContext context,
-    int minutes,
+    double minutes,
     AppLocalizations localizations,
   ) {
     Navigator.pop(context);
 
     final playerService = getIt<GlobalPlayerService>();
-    playerService.setSleepTimer(minutes);
+    playerService.setSleepTimer(minutes.toInt());
 
     ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(content: Text(localizations.timerSetMessage(minutes))),
+      SnackBar(content: Text(localizations.timerSetMessage(minutes.toInt()))),
     );
 
     onTimerSet?.call();
